Move overrideGenerationTemplate function to utils.ts
